Backend errors array  and blogs , offers  auto refresh added

+++ backend/src/controller/products.ts
@@ -57,12 +57,11 @@ export async function createProduct(
     });
 
     const offersRepository = getConnection().getRepository(Offer);
-    const offer = await offersRepository.findOne({ id: offerId });
 
     const productImageRepository = getConnection().getRepository(ProductImage);
-
+    let offer;
     if (offerId) {
-      let offer = await offersRepository.findOne({ id: offerId });
+      offer = await offersRepository.findOne({ id: offerId });
     }
 
     const createProductImage = [];
@@ -80,7 +79,7 @@ export async function createProduct(
         createProductImage.push(savedProductImage);
       }
     } else {
-      return res.json("Image not found");
+      return res.json({ errors: [{ msg: "Image not found" }] });
     }
 
     const newProduct = new Product();
@@ -92,7 +91,7 @@ export async function createProduct(
     if (categoryCheck) {
       newProduct.category = categoryId;
     } else {
-      res.status(400).json({ msg: "category not found" });
+      res.status(400).json({ errors: [{ msg: "category not found" }] });
     }
 
     if (offer) {
@@ -102,7 +101,7 @@ export async function createProduct(
     // save data to repository from request body
     await productsRepository.save(newProduct);
   } catch (e) {
-    res.status(400).json({ error: e });
+    res.status(400).json({ errors: [{ msg: e }] });
     return;
   }
   res.json({ msg: "Product created" });
@@ -117,7 +116,7 @@ export async function getProduct(req: express.Request, res: express.Response) {
   const findProductById = await productRepository.findOne({ id });
 
   if (!findProductById) {
-    return res.status(400).json({ error: "Product not found" });
+    return res.status(400).json({ errors: [{ msg: "Product not found" }] });
   }
 
   res.json({ msg: "product found", data: findProductById });
@@ -154,7 +153,9 @@ export async function updateProduct(
 
     await productsRepository.save(findProductById);
   } catch (e) {
-    return res.status(400).json({ error: "Product could not be updated" });
+    return res
+      .status(400)
+      .json({ errors: [{ msg: "Product could not be updated" }] });
   }
 
   res.json({ msg: "Product updated" });
@@ -168,11 +169,15 @@ export async function deleteProduct(
 ) {
   const id = req.params.productId;
   const productRepository = getConnection().getRepository(Product);
-
-  try {
-    await productRepository.delete(id);
-  } catch (e) {
-    return res.status(400).json({ error: "Product could not be deleted" });
+  const productCheck = await productRepository.findOne({ id });
+  if (productCheck) {
+    try {
+      await productRepository.delete(id);
+    } catch (e) {
+      return res
+        .status(400)
+        .json({ errors: [{ msg: "Product could not be deleted" }] });
+    }
+    res.json({ msg: "Product deleted" });
   }
-  res.json({ msg: "Product deleted" });
 }
